requirements.txt
gunicorn

# Web framework
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.0.0

# LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Basic utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0

# Additional dependencies for RAG
pinecone-client>=3.0.0
langchain-pinecone>=0.0.3
PyPDF2>=3.0.0

# OpenAI for embeddings and LLM
langchain-google-genai>=0.0.3
